+++ var/cache/templates/ee/ee5b7fc0a4d87dd2e944f30710370c4b7850f172a35246d53ee2476696857cee.php
@@ -1,0 +1,97 @@
+<?php
+
+use Twig\Environment;
+use Twig\Error\LoaderError;
+use Twig\Error\RuntimeError;
+use Twig\Extension\SandboxExtension;
+use Twig\Markup;
+use Twig\Sandbox\SecurityError;
+use Twig\Sandbox\SecurityNotAllowedTagError;
+use Twig\Sandbox\SecurityNotAllowedFilterError;
+use Twig\Sandbox\SecurityNotAllowedFunctionError;
+use Twig\Source;
+use Twig\Template;
+
+/* error.html.twig */
+class __TwigTemplate_706bd9cef7eafc287f544891fe451be71706a429643cbbee61a92946338d8e4d extends Template
+{
+    private $source;
+    private $macros = [];
+
+    public function __construct(Environment $env)
+    {
+        parent::__construct($env);
+
+        $this->source = $this->getSourceContext();
+
+        $this->blocks = [
+            'title' => [$this, 'block_title'],
+            'body' => [$this, 'block_body'],
+            'bottom_js' => [$this, 'block_bottom_js'],
+        ];
+    }
+
+    protected function doGetParent(array $context)
+    {
+        // line 1
+        return "layout.html.twig";
+    }
+
+    protected function doDisplay(array $context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        $this->parent = $this->loadTemplate("layout.html.twig", "error.html.twig", 1);
+        $this->parent->display($context, array_merge($this->blocks, $blocks));
+    }
+
+    // line 3
+    public function block_title($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        echo " Erreur ";
+    }
+
+    // line 5
+    public function block_body($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        // line 6
+        echo "<h1>Bonjour il semblerait qu'il y ait eu une erreur</h1>
+
+<p>";
+        // line 8
+        echo twig_escape_filter($this->env, ($context["error"] ?? null), "html", null, true);
+        echo "<p>
+
+";
+    }
+
+    // line 12
+    public function block_bottom_js($context, array $blocks = [])
+    {
+        $macros = $this->macros;
+        // line 13
+        echo "       
+";
+    }
+
+    public function getTemplateName()
+    {
+        return "error.html.twig";
+    }
+
+    public function isTraitable()
+    {
+        return false;
+    }
+
+    public function getDebugInfo()
+    {
+        return array (  74 => 13,  70 => 12,  63 => 8,  59 => 6,  55 => 5,  48 => 3,  37 => 1,);
+    }
+
+    public function getSourceContext()
+    {
+        return new Source("", "error.html.twig", "/home/valentin/Documents/php_proj/board-game/templates/error.html.twig");
+    }
+}