+++ CHANGELOG.md
@@ -4,6 +4,16 @@ Tous les changements de ce projet seront documentés dans ce fichier.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+
+## 04-12-2021
+## Added
+- Ajout class Modele pour la serialisations
+
+## Changed
+- Attribut des classes métier passés de private a protected
+
+
 ## 02-12-2021
 ## Added
 - Ajout design bouton et inpu